Rename README and add Build directory
This commit amended the previous commit by adding the Build directory
and gives a multiple line comment. To allow 'git commit' to work, the
git core editor must be 'gvim -f', not just 'gvim'. This means gvim will
run in the foreground and will not give control back to git immediately
once it begins. This way, you can enter in a longer commit message
instead of having to use 'git commit -m' all of the time.

- The is a bullet point.