Create the LinkGitHooks.sh script to link git hooks to .git/hooks
Since a cloned repository does not carry the hooks used in a local repository,
the LinkGitHooks.sh script was made to link the hooks stored in the repository
to the .git/hooks directory. This script is executed in the GramSchmidt.sh
script when a "setenv" argument is passed. It can also be executed by itself.

File: Build/GramSchmidt.sh
--------------------------
- The LinkGitHooks.sh script is being run in the "setenv" conditional
- Status messages were added in the "setenv" conditional to let the user know
  what the script is doing as it executes

File: hooks/commit-msg
----------------------
- Changes were made to handle the case when no files are mentioned in
  the commit message
- The "Files not staged for commit" error message was moved to the
  print_msgFilesErr function

File: hooks/prepare-commit-msg
------------------------------
- Removed the ":" from the warning message if the -m or -F arguments are used

File: Tools/Scripts/LinkGitHooks.sh
-----------------------------------
- Link the repository git hooks to the .git/hooks directory, if they exist