Revised files to NOT use the clean filter
Since the clean filter is not able to expand keywords in a manner that would
cause no errors to occur when renormalizing files in a merge, it was decided to
not use these filters and do away with trying to perform keyword expansion.

Instead, expansion of Format: fields will be used when creating an archive and
file properties such as File, Author, and Date Revised will be updated manually.

File: .gitattributes
--------------------
- Replaced the filter attribute with the export-subst attribute when creating
  archives for release

File: .gitignore
----------------
- Updated comments back to what they were before testing the clean filter

File: Tools/Scripts/GitKeywordExp.sh
------------------------------------
- Removed this file, since the clean filter is not longer being used