Allow merge commit messages to be accepted as is
When merge conflicts occur, the user must commit the changes made to the
conflicting files after the conflicts are resolved. This results in the
commit-msg hook running and performing format checks on the commit message. This
is often undesirable, since unnecessary formatting errors can occur. To bypass
this default behavior, the commit-msg hook prompts the user for whether or not
they would like to accept the message as is or continue with the regular format
checking if a merge commit is detected.

Merge commits are detected if the first line in the commit message contains the
default "Merge branch '<branch name>'" message or if the .git/MERGE_HEAD file
exists.

File: hooks/commit-msg
----------------------
- Detects if the commit is for a merge conflict and prompts the user for which
  action to take