Fixed Makefile compilation error for application source files
It was discovered that the way the "all" target was reinvoking Make with a new
target, it reset the MAKECMDGOALS variable. This variable was being used to
determine when the dependency files should be included and was only including
them when MAKECMDGOALS was equal to "all". This was not occuring when the
executable needed to be built, which was resulting in the GramSchmidt.o file
only depending on the GramSchmidt.cc file and no others. This meant that if a
change was made in an included header file, the GramSchmidt.o file would not be
updated and the compiled program would not see the code change.

This has been fixed with a change to the application "all" target, which now has
the application target as a prerequisite, instead of reinvoking Make to compile
it.

File: Build/StdMake/Makefile.app
--------------------------------
- Deleted the recipe for the "all" target
- Added a new "local_all" prerequisite
- Added a new "local_configure" prerequisite to match the Makefile.libs usage

File: GramSchmidt_App/src/GramSchmidt.cc
----------------------------------------
- Minor code changes while getting the Vector class to work

File: GramSchmidt_App/src/Makefile
----------------------------------
- Added the local_all and local_configure targets and prerequisites

File: Utilities/libutlmath/header/Vector.hh
-------------------------------------------
- Minor code changes while getting the Vector class to work

File: Utilities/libutlmath/src/Vector.cc
----------------------------------------
- Minor code changes while getting the Vector class to work